refactor: aplicando regras de negocio a funcionarios, e retirando excesso de regras da controller

+++ src/main/java/edu/davi/api/tarefa/service/FuncionarioService.java
@@ -1,13 +1,14 @@
 package edu.davi.api.tarefa.service;
 
-import java.util.List;
-import java.util.Optional;
-
+import edu.davi.api.tarefa.exception.BusinessException;
+import edu.davi.api.tarefa.model.Funcionario;
+import edu.davi.api.tarefa.repository.FuncionarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import edu.davi.api.tarefa.model.Funcionario;
-import edu.davi.api.tarefa.repository.FuncionarioRepository;
+import javax.persistence.EntityNotFoundException;
+import java.util.List;
+import java.util.Optional;
 
 @Service
 public class FuncionarioService {
@@ -21,35 +22,56 @@ public class FuncionarioService {
 
 	public List<Funcionario> listarTodos() {
 
-		return funcionarioRepository.findAll();
+		List<Funcionario> result = funcionarioRepository.findAll();
 
+		if (result.isEmpty()) {
+			throw new BusinessException("Não existem funcionários cadastrados");
+		}
+		return result;
 	}
 
 	public Optional<Funcionario> listarPorId(Long id) {
+		Optional<Funcionario> funcionario = funcionarioRepository.findById(id);
 
-		return funcionarioRepository.findById(id);
-
+		if (funcionario.isEmpty()) {
+			throw new EntityNotFoundException("Funcionário não encontrado");
+		}
+		return funcionario;
 	}
 
 	public Optional<Funcionario> findByCpfFuncionario(String cpfFuncionario) {
-		return funcionarioRepository.findByCpfFuncionario(cpfFuncionario);
+		Optional<Funcionario> funcionario = funcionarioRepository.findByCpfFuncionario(cpfFuncionario);
+		if (funcionario.isEmpty()) {
+			throw new EntityNotFoundException("CPF não encontrado");
+		}
+		return funcionario;
 	}
 
 	public Funcionario salvar(Funcionario funcionario) {
-
+		if (findByCpfFuncionario(funcionario.getCpfFuncionario()).isPresent()) {
+			throw new BusinessException("CPF já cadastrado");
+		} else if (funcionario.getCpfFuncionario().isEmpty())
+			throw new BusinessException("CPF é obrigatório");
 		return funcionarioRepository.save(funcionario);
-
 	}
 
 	public void deletar(Long id) {
-		funcionarioRepository.deleteById(id);
+		if (listarPorId(id).isPresent())
+			funcionarioRepository.deleteById(id);
+		throw new IllegalStateException("Não foi possivel deletar");
 	}
 
 	public Funcionario alterar(Long id, Funcionario funcionario) {
 
-		funcionario.setId(id);
-		return funcionarioRepository.save(funcionario);
+		Optional<Funcionario> funcionarioParaEditar = listarPorId(id);
 
+		if (funcionarioParaEditar.isPresent()) {
+			funcionarioParaEditar.get().setNomeFuncionario(funcionario.getNomeFuncionario());
+			funcionarioParaEditar.get().setCpfFuncionario(funcionario.getCpfFuncionario());
+			funcionarioParaEditar.get().setDataNasc(funcionario.getDataNasc());
+			return funcionarioRepository.save(funcionarioParaEditar.get());
+		}
+		throw new BusinessException("Falha ao editar funcionario");
 	}
 
 }
